Feature/test and fix

+++ src/main/java/com/neoris/user_flow_api/delegate/impl/AccountDelegateImpl.java
@@ -1,5 +1,6 @@
 package com.neoris.user_flow_api.delegate.impl;
 
+import com.neoris.user_flow_api.constans.NotificationCode;
 import com.neoris.user_flow_api.delegate.AccountDelegate;
 import com.neoris.user_flow_api.domain.Account;
 import com.neoris.user_flow_api.domain.Customer;
@@ -36,9 +37,11 @@ public class AccountDelegateImpl implements AccountDelegate {
     for (AccountDTO accountDTO : accountDTOs) {
       ValidationUtils.validateAccountType(accountDTO);
       ValidationUtils.validateValue(accountDTO.getBalance());
-      ValidationUtils.validateAccountType(accountDTO);
       Customer customer = customerService.findById(accountDTO.getCustomerId());
-
+      Account accountFound = accountService.findById(accountDTO.getAccountNumber());
+      if (accountFound != null) {
+        throw new UserFlowException(NotificationCode.ACCOUNT_ALREADY_EXISTS);
+      }
       Account account = AccountMapper.INSTANCE.accountDTOToAccount(accountDTO);
       account.setCustomer(customer);
       accountsToSaved.add(account);
@@ -60,6 +63,10 @@ public class AccountDelegateImpl implements AccountDelegate {
 
     Account accountFound = accountService.findById(id);
 
+    if (accountFound == null) {
+      throw new UserFlowException(NotificationCode.ACCOUNT_NOT_FOUND);
+    }
+
     AccountMapper.INSTANCE.updateAccountFromAccountDTO(accountDTO, accountFound);
     Account updatedAccount = accountService.save(accountFound);
     log.info("{} account updated", id);
@@ -70,12 +77,18 @@ public class AccountDelegateImpl implements AccountDelegate {
   @Override
   public AccountDTO getAccountById(Long accountNumber) throws UserFlowException {
     Account account = accountService.findById(accountNumber);
+    if (account == null) {
+      throw new UserFlowException(NotificationCode.ACCOUNT_NOT_FOUND);
+    }
     return AccountMapper.INSTANCE.accountToAccountDTO(account);
   }
 
   @Override
   public void deleteAccount(Long id) throws UserFlowException {
     Account account = accountService.findById(id);
+    if (account == null) {
+      throw new UserFlowException(NotificationCode.ACCOUNT_NOT_FOUND);
+    }
     accountService.delete(account);
     log.info("{} account deleted", id);
   }
